Bug 568117: Handle case where there was no staging repo already
After a release, the staging repo is gone so it can't be moved
out of the way. Handle that case with this patch.

+++ releng/org.eclipse.epp.config/tools/upload-to-staging.sh
@@ -90,7 +90,9 @@ done
 ${SSH} rm -rf ${STAGING}-new
 ${SCP} -rp . "${SSHUSER}:"${STAGING}-new
 ${SSH} rm -rf ${STAGING}-previous
-${SSH} mv ${STAGING} ${STAGING}-previous
+if $SSH test -e ${STAGING}; then
+  ${SSH} mv ${STAGING} ${STAGING}-previous
+fi
 ${SSH} mv ${STAGING}-new ${STAGING}
 ${SSH} rm -rf ${STAGING}-previous
